fix(vae): address cpu_only VAE review feedback (#9293)
Item 1 — VAE settings panel showed text-encoder-specific copy:
- Add `runVaeOnCpu` label and a `cpuOnlyVae` popover feature with
  VAE-accurate wording instead of reusing the encoder's `runOnCpu` /
  `cpuOnly` strings (which describe the text encoder and conditioning
  outputs, wrong on the VAE panel).
- Register `cpuOnlyVae` in the InformationalPopover feature union.

Item 2 — mislabeled INFO log on GPU-less installs:
- ModelCache.lock only logs the "(cpu_only)" cause at INFO when the
  cache's default execution device is a GPU (a genuine per-model
  choice). On a CPU-only install every model has a CPU compute_device,
  so that case now logs at DEBUG without the misleading wording.

Item 3 — stale l2i comment:
- Replace the false "fp16 VAE ops are not supported on CPU" rationale
  with the real reason: fp16 conv runs on CPU with the pinned torch but
  is much slower, and SD/SDXL VAE has known fp16 overflow (black
  images). The fp32 force is unchanged.

Item 4 — frontend duplication:
- Extract the shared CPU-only settings panel into CpuOnlyModelSettings
  and a single useCpuOnlyModelSettings hook, parameterized by label,
  popover feature, and toast id. EncoderModelSettings and
  VAEModelSettings become thin wrappers; the config unions stay
  separate and each panel keeps its own copy. Removes ~200 lines of
  duplicated code and the awkward cross-folder type import; deletes the
  now-redundant useEncoderModelSettings / useVAEModelSettings hooks.

SD3 / CogView4 left unchanged (out of scope):
- No standalone VAE config exists for these bases, and a main model's
  default_settings.cpu_only only applies to text-encoder submodels
  (load_default.py), so cpu_only can never reach an SD3/CogView4 VAE —
  their l2i changes are harmless consistency, nothing to toggle.
- Adding standalone VAE configs for them is not worth it here: there is
  no real ecosystem of separate SD3/CogView4 VAE files, and their VAEs
  are 16-channel AutoencoderKL that can't be reliably told apart from
  each other (or Flux) at import-probe time — which is exactly why
  standalone VAE configs are limited to the distinguishable bases
  (SD1/SD2/SDXL, Flux). The runtime is already cpu_only-ready, so if
  such a config is ever added the toggle works with no further change.
