+++ invokeai/frontend/web/src/common/components/InformationalPopover/constants.ts
@@ -78,6 +78,7 @@ export type Feature =
   | 'optimizedDenoising'
   | 'fluxDevLicense'
   | 'cpuOnly'
+  | 'cpuOnlyVae'
   | 'fp8Storage';
 
 export type PopoverData = PopoverProps & {
